add trait FileSourceReader
